% Generated by roxygen2 (4.0.0): do not edit by hand
\name{addError}
\alias{addError}
\title{Non-fatal error tracking}
\usage{
addError(errorMessage, errorEnv = NULL)
}
\arguments{
\item{errorMessage}{The message for the user}

\item{errorEnv}{If the errorEnv is not provided, the search path is used to find an errorList.
Providing an errorEnv is the preferred method to avoid name collsions.}
}
\description{
Add errors to a list of errors to show a user
}

