% Generated by roxygen2 (4.0.0): do not edit by hand
\name{addFileLink}
\alias{addFileLink}
\title{Saves a file that will show up as a link in Seurat}
\usage{
addFileLink(batchCodeList, recordedBy, experiment, lsTransaction,
  reportFileSummary = NULL, reportFilePath = NULL, fileType = NULL,
  url = NULL, testMode = FALSE)
}
\arguments{
  \item{batchCodeList}{A list or vector of batch codes}

  \item{recordedBy}{The username of the person recording
  the data}

  \item{experiment}{A list that is an experiment object,
  needs to have values for id, verison, and codeName}

  \item{lsTransaction}{A number that is a transaction id}

  \item{reportFileSummary}{A description of the file or url
  annotation}

  \item{reportFilePath}{The path to the report file (may be
  relative)}

  \item{url}{A url (e.g. http://google.com)}

  \item{fileType}{The type of the annotation, used by some
  custom services}

  \item{testMode}{a boolean used for testing}
}
\description{
Saves an analysis group with states and values holding a
file or url. Each batchCode has a separate state with the
annotation.
}
\details{
Must have either a reportFilePath or a url, but having both
will result in the url being ignored.
}

