% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tsvSave.R
\name{analysis_group_save_from_tsv}
\alias{analysis_group_save_from_tsv}
\title{Save file paths using tsv loader}
\usage{
analysis_group_save_from_tsv(analysisGroupCsvFilePath,
  treatmentGroupCsvFilePath = NULL, subjectCsvFilePath = NULL)
}
\arguments{
\item{analysisGroupCsvFilePath}{A file path to analysis group values tsv file}

\item{optional}{treatmentGroupCsvFilePath A file path to treatment group values tsv file}

\item{optional}{subjectCsvFilePath A file path to subject group values tsv file}
}
\description{
Saves tsv files all in one transaction
}
\details{
files must be a full path reachable to the acas roo services
}

