\name{api_createProtocol}
\alias{api_createProtocol}
\title{Create acas protocol}
\usage{
  api_createProtocol(protocolName, shortDescription,
    recordedBy, stopOnAnyExist = TRUE)

  api_createProtocol(protocolName, shortDescription,
    recordedBy, stopOnAnyExist = TRUE)
}
\arguments{
  \item{protocolName}{}

  \item{shortDescription}{}

  \item{protocolName}{}

  \item{shortDescription}{}
}
\value{
  recordedBy

  recordedBy
}
\description{
  This function creates an acas protocol

  This function creates an acas protocol
}
\examples{
api_createProtocol(c("Brian Test Protocol2"), shortDescription = "Test Protocol", recordedBy = "bbolt")
api_createProtocol(c("Brian Test Protocol2"), shortDescription = "Test Protocol", recordedBy = "bbolt")
}
\keyword{create,}
\keyword{protocol}
\keyword{register,}

