% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dose_response_api.R
\name{api_doseResponse_experiment}
\alias{api_doseResponse_experiment}
\title{Fits experiment dose response data}
\usage{
api_doseResponse_experiment(simpleFitSettings, modelFitType, recordedBy,
  experimentCode, testMode = NULL, modelFit)
}
\arguments{
\item{simpleFitSettings}{list of fit settings (see details)}

\item{recordedBy}{character user that curve data will be saved as}

\item{experimentCode}{character saved experiment code that contains dose response data}

\item{testMode}{logical unimplemented}
}
\value{
a list of the entities returned from the server
}
\description{
This function retrieves the dose response data for a given experiment code, refits it using the simple settings and then saves the data back to the database
}
\examples{
file <- system.file("docs", "example-simple-fitsettings-ll4.json", package = "racas" )
simpleBulkDoseResponseFitRequestJSON <- readChar(file, file.info(file)$size)
simpleFitSettings <- fromJSON(simpleBulkDoseResponseFitRequestJSON)
recordedBy <- "bbolt"
experimentCode <- "EXPT-00000441"
modelFitType <- "4 parameter D-R"
api_doseResponse_experiment(simpleFitSettings, modelFitType, recordedBy, experimentCode)

#Loading fake data first
# requires 1. that a protocol named "Target Y binding") be saved first (see \\code{\\link{api_createProtocol}})
#          2. have a valid recordedBy (or that acas is set to client.require.login=false)

file <- system.file("docs", "example-simple-fitsettings-ll4.json", package = "racas" )
simpleBulkDoseResponseFitRequestJSON <- readChar(file, file.info(file)$size)
simpleFitSettings <- fromJSON(simpleBulkDoseResponseFitRequestJSON)
experimentCode <- load_dose_response_test_data()
recordedBy <- "bbolt"
api_doseResponse_experiment(simpleFitSettings, recordedBy, experimentCode)
}
\keyword{dose,}
\keyword{experimet}
\keyword{fit,}
\keyword{model,}
\keyword{response,}

