% Generated by roxygen2 (4.0.2): do not edit by hand
\name{api_doseResponse_fitData_to_curveDetail}
\alias{api_doseResponse_fitData_to_curveDetail}
\title{fitData object to json response}
\usage{
api_doseResponse_fitData_to_curveDetail(fitData, saved = TRUE, ...)
}
\arguments{
\item{...}{addition arguments to be passed to \code{\link{toJSON}}}

\item{a}{fitData object}
}
\value{
A json object of the fitData and any other objects coerced to json by ... \code{\link{toJSON}}
}
\description{
Converts a fitData object to a json response to return to the ACAS GUI
}
\examples{
#Load and example fitData object
data("example-ec50-fitData-fitted")
#FitData object plus the "cars" data to a json string
api_doseResponse_fitData_to_curveDetail(fitData, cars)
}

