% Generated by roxygen2 (4.0.2): do not edit by hand
\name{appendCodeNames}
\alias{appendCodeNames}
\title{Appends code names to supplied valueKinds}
\usage{
appendCodeNames(entityData, valueKinds, entitySpaced)
}
\arguments{
\item{entityData}{data.frame that will be changed}

\item{valueKinds}{vector of valueKinds to have stringValues changed}

\item{entitySpaced}{the name of the entity: "analysis group" or "treatment
  group" or "subject"}
}
\description{
Appends code names to supplied valueKinds
}
\details{
not for export, used only by saveAllViaTsv. Supports data.table and
  data.frame. In tsvSave.R
}

