% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{applicationSettings}
\alias{applicationSettings}
\title{Configuration settings for the applicaiton}
\format{A data frame with a column for each setting}
\usage{
applicationSettings
}
\description{
This data frame is used throughout the applicaiton to point to the database, to get the correct applicaiton name...etc.
}
\details{
\itemize{
  \item appName display name for the applicaiton (ACAS)
  \item db_driver a character string that when evaluated is an object that inherits from \code{\link{dbDriver}} ('PostgreSQL()', 'Oracle()', 'MySQL()', 'JDBC("oracle.jdbc.driver.OracleDriver", "/Users/bbolt/Documents/ojdbc6-11.2.0.2.jar")')
  \item db_user a character string username for the database connection
  \item db_password a character string password for the database connection
  \item db_name a character string SID/database name for the database connection
  \item db_host a character string hostname of database connection
  \item db_port a character string port of database connection
  \item db_driver_package a character string of the package to instantiate
}
}
\examples{
# Here is an example data frame
applicationSettings <- data.frame(
appName = "ACAS",               #Application Display Name
db_driver = "PostgreSQL()",     #Must be supplied in your own package load (MySQL(), Oracle() supported)
db_user = "username",           #ACAS Schema db user
db_password = "password",       #ACAS Schema db password
db_name = "databasename",       #ACAS Schema db name
db_host = "mcneilco.com",       #ACAS Host Name
db_port = "5432",               #ACAS Port Number
stringsAsFactors = FALSE
)
# You must assign it to the package namespace in order for it to be used by the package
assignInNamespace("applicationSettings",applicationSettings, ns="racas")
}
\keyword{datasets}

