% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JSON_library.R
\name{createCodeTablesFromJsonArray}
\alias{createCodeTablesFromJsonArray}
\title{createCodeTablesFromJsonArray}
\usage{
createCodeTablesFromJsonArray(codeTableDataFrame,
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{codeTableDataFrame}{of ddict values}
}
\value{
a data frame of kinds
}
\description{
Create code table (d dict values) from json array as described here RCurl::curlEscape(paste0(racas::applicationSettings$client.service.persistence.fullpath,"ddictvalues/codetable/jsonArray"))
}

