% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{createSELCalculatedResults}
\alias{createSELCalculatedResults}
\title{Creates a data.frame that contains the elements of the Calculated Results section of SEL}
\usage{
createSELCalculatedResults(calculatedResults,
  corporateIDColumn = "Corporate Batch ID", dataTypes = NULL,
  ls_kinds = NULL, unit_kinds = NULL)
}
\arguments{
  \item{calculatedResultDataFrame}{a data.frame (or
  data.table) of calculated results}

  \item{dataTypes}{optional character with length equal to
  ncol(calculatedResultDataFrame) with the data types for
  each column}

  \item{ls_kinds}{optional character with length equal to
  ncol(calculatedResultDataFrame) with the ls_kinds for
  each column}

  \item{unit_kinds}{optional character with length equal to
  ncol(calculatedResultDataFrame) with the unit_kinds for
  each column}
}
\value{
data.frame representing the most common inputs to the SEL
calculated results section
}
\description{
The return structure represents the standard inputs of the
"Calculated Results" section of SEL
}
\examples{
#Basic Example
cars2 <- cars
cars2$Sample <- paste0("CORP",1:nrow(cars))
createSELCalculatedResults(cars2, corporateIDColumn = "Sample")

#This gives warning that "Corporte Batch ID" not found
createSELCalculatedResults(cars)

#Additional Features
createSELCalculatedResults(cars2, corporateIDColumn = "Sample", dataTypes = c("Number","Text"))
createSELCalculatedResults(cars2, corporateIDColumn = "Sample", unit_kinds = c("miles per hour", "kilometers"))
createSELCalculatedResults(cars2, corporateIDColumn = "Sample", ls_kinds = c("IC50", "Max Efficacy"), unit_kinds = c("uM", "\%"))
createSELCalculatedResults(cars2, corporateIDColumn = "Sample", ls_kinds = c("IC50 (still includes this)", "Max Efficacy"), unit_kinds = c("uM", "\%"))

#Error handling
createSELCalculatedResults(cars2, corporateIDColumn = "Some Garbage")
createSELCalculatedResults(cars2, corporateIDColumn = "Sample", dataTypes = "Some Garbage")
}

