% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sel.R
\name{createSELFile}
\alias{createSELFile}
\title{Creates an SEL formatted file}
\usage{
createSELFile(selExperimentMetaData, selCalculatedResults = NULL,
  selRawResults = NULL, outputFilePath = NULL, format = "CSV")
}
\arguments{
\item{selExperimentMetaData}{typically a data.frame created by \code{\link{createSELExperimentMetaData}}}

\item{selCalculatedResults}{optional data.frame of the format \code{\link{createSELCalculatedResults}}}

\item{selRawResults}{optional data.frame of the format \code{\link{createSELRawResults}}}

\item{outputFilePath}{optional output file path (see details for return if left NULL)}

\item{format}{"CSV", "XLS" or "XLSX" (not used if outputFilePath is NULL, see details)}
}
\value{
character path of the written file or character of the CSV results
}
\description{
Given data structures that represent the various pieces of an SEL file, this function will write or return SEL formatted files
}
\details{
If the outputFilePath is NULL then the function will write the CSV representation of the file to a character and return, otherwise it returns the path to the written file.
}
\examples{
selExperimentMetaData <- createSELExperimentMetaData(protocolName = "Cytotox Assay",
                                                experimentName = "2013123_cytotox_bbolt",
                                                assayDate = "2013-01-12")

cars2 <- cars
cars2$Sample <- paste0("CORP",1:nrow(cars))
selCalculatedResults <- createSELCalculatedResults(cars2, corporateIDColumn = "Sample")

#Outputs a character representation of a CSV SEL format
createSELFile(selExperimentMetaData, selCalculatedResults)

#Outputs a file in SEL format
myFile <- tempfile()
createSELFile(selExperimentMetaData, selCalculatedResults, outputFilePath = myFile)
read.csv(myFile)
}

