\docType{data}
\name{createSELRawResults}
\alias{createSELRawResults}
\title{Creates a data.frame that contains the elements of the Raw Results section of SEL}
\usage{
  createSELRawResults(rawResults,
    curveIDColumn = "curve id", ls_kinds = NULL,
    unit_kinds = NULL)

  createSELRawResults
}
\description{
  The return structure represents the standard inputs of
  the "Raw Results" section of SEL
}
\examples{
Not implemented yet
}
\keyword{datasets}

