% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/directDatabaseSave.R
\name{dbWriteTableMatchCol}
\alias{dbWriteTableMatchCol}
\title{Write table with matching columns names}
\usage{
dbWriteTableMatchCol(conn, name, value, ...)
}
\arguments{
\item{conn}{a DBIConnection.}

\item{name}{A character string specifying a DBMS table name.}

\item{value}{a data.table to write.}
}
\description{
Writes a data.table to a database table like \code{\link{dbWriteTable}}, but
matches to the column names. Column names must be all lowercase in your
data.table.
}

