% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dose_response.R
\name{dose_response}
\alias{dose_response}
\title{Fit dose response data}
\usage{
dose_response(fitSettings, fitData)
}
\arguments{
\item{fitData}{a fitData object to fit/refit}

\item{fitSettingsJSON}{a fit settings json object (see examples)}
}
\value{
A fitted fitData object
}
\description{
Converts a fitData object to a fitted fitData object (adds model, reported parameters...etc.)
}
\examples{
#Fit all the curveids in the database:

#get fitData
data("example-ec50-fitData")

#fitSettingsJSON
file <- system.file("conf", "default-ec50-fitSettings.json", package = "racas")
fitSettingsJSON <- readChar(file, file.info(file)$size)
fitSettings <- fromJSON(fitSettingsJSON)
#fit the data
fitData <- dose_response(fitSettings, fitData)
}

