% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dose_response_session}
\alias{dose_response_session}
\title{Fit dose response data}
\usage{
dose_response_session(fitSettings, curveids = NA, sessionID = NA,
  fitData = NA, simpleFitSettings = NULL, flagUser = NULL, user = NULL,
  ...)
}
\arguments{
\item{fitSettings}{a fit settings list object (see examples)}

\item{curveids}{a character list of curveids}

\item{sessionID}{a path to a curve fit session}

\item{fitData}{a fidData object to refit}

\item{simpleFitSettings}{a list of simplefitsettings to attach to the fitData object}
}
\value{
A list object with a fitted fitData object and sessionID
}
\description{
Converts a character vector of curveids, character sessionsID or data.table fitData object to a fitted fitData object,
saves the fitData object to either a new session or back to the session provided and then returns
a json representation of the fitted fitData object with the savedSessionID
}
\examples{
#Fit all the curveids in the database:

#get curveids
curveids <- as.character(query("select curveid from api_curve_params")[[1]])

#fitSettings
file <- system.file("conf", "default-ec50-fitSettings.json", package = "racas")
fitSettingsJSON <- readChar(file, file.info(file)$size)
fitSettings <- fromJSON(fitSettingsJSON)
#fit the data
system.time(response <- get_fit_data_experiment_codefitSettings, curveids = curveids))
}

