% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dose_response.R
\name{fit_data_to_acas_experiment_response}
\alias{fit_data_to_acas_experiment_response}
\title{fitData object to json response}
\usage{
fit_data_to_acas_experiment_response(fitData, experimentCode,
  transactionId = -1, status, hasWarning, errorMessages = as.character(),
  ...)
}
\arguments{
\item{...}{addition arguments to be passed to \code{\link{toJSON}}}

\item{a}{fitData object}
}
\value{
A json object of the fitData and any other objects coerced to json by ... \code{\link{toJSON}}
}
\description{
Converts a fitData object to a json response to return to the GUI
}
\examples{
#Load and example fitData object
data("example-ec50-fitData-fitted")
#FitData object plus the "cars" data to a json string
fit_data_to_acas_experiment_response(fitData, cars)
}

