% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JSON_library.R
\name{flattenLabel}
\alias{flattenLabel}
\title{Flattens an lsLabel}
\usage{
flattenLabel(lsLabel)
}
\arguments{
\item{lsLabel}{an lsLabel

Just turns a list into a data frame, not meant to be exported}
}
\description{
Flattens an lsLabel
}

