% Generated by roxygen2 (4.0.2): do not edit by hand
\name{flattenLabels}
\alias{flattenLabels}
\title{Flattens a list of lsLabels}
\usage{
flattenLabels(lsLabels)
}
\arguments{
\item{lsLabels}{a list os lsLabels

Just turns a list into a data frame, not meant to be exported}
}
\description{
Flattens a list of lsLabels
}

