% Generated by roxygen2 (4.0.1): do not edit by hand
\name{flattenState}
\alias{flattenState}
\title{Flattens an lsState}
\usage{
flattenState(lsState, includeFromState)
}
\arguments{
  \item{lsState}{an lsState that has lsValues}

  \item{includeFromState}{a character vector of column
  names to include from the state

  Will return an empty data frame if there are no lsValues}
}
\description{
Gets values into a data.frame
}

