% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JSON_library.R
\name{flattenValue}
\alias{flattenValue}
\title{Flattens an lsValue}
\usage{
flattenValue(lsValue)
}
\arguments{
\item{lsValue}{an lsValue

Just turns a list into a data frame, not meant to be exported}
}
\description{
Flattens an lsValue
}

