% Generated by roxygen2 (4.0.1): do not edit by hand
\name{formatEntityAsTsvAndUpload}
\alias{formatEntityAsTsvAndUpload}
\title{Save data using tsv}
\usage{
formatEntityAsTsvAndUpload(entityData)
}
\arguments{
\item{entityData}{data.frame with set of columns- see code for list}
}
\value{
path to file on persistence server
}
\description{
Saves data frames to a tsv and uploads to persistence server
}
\details{
Assumes all data is new, not saving to existing entities
}

