% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getAcasFileLink}
\alias{getAcasFileLink}
\title{Get url for ACAS file}
\usage{
getAcasFileLink(fileCode)
}
\arguments{
\item{fileCode}{A file code of some custom type like "FILE1234" or a path
link like "experiment/EXPT-3/this.txt"}
}
\value{
a url
}
\description{
Gets a url from an ACAS filecode. This deals with issues of whether the file
is stored internally or on an external file system, so you just get a link to
wherever the file is stored.
}
\details{
The getting equivalent of \code{\link{saveAcasFile}}. In
  updateExperimentMetadata.R
}

