% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JSON_library.R
\name{getAllValueKinds}
\alias{getAllValueKinds}
\title{valueKinds}
\usage{

  getAllValueKinds(lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{lsServerURL}{url for roo server}
}
\description{
Gets a list of all valueKinds available.
}

