% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JSON_library.R
\name{getAllValueTypes}
\alias{getAllValueTypes}
\title{valueTypes}
\usage{

  getAllValueTypes(lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{lsServerURL}{url for roo server}
}
\description{
Gets a list of all valueTypes available.
}

