% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preferredBatchCode.R
\name{getCompoundName}
\alias{getBatchNumber}
\alias{getCompoundName}
\title{Gets parts of batch code}
\usage{
getCompoundName(batchCode)

getBatchNumber(batchCode)
}
\arguments{
\item{batchCode}{vector of batch codes}
}
\value{
character vector
}
\description{
Takes a batch code and splits it on the last batch separator, as defined in
\code{racas::applicationSettings$server.service.external.preferred.batchid.separator}
getCompoundName will return everything except the last section, while
getBatchName will return everything after the last separator. So "CMPD-000001-01A"
is split into "CMPD-000001" and "01A" when the separator is "-"
}

