% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entityMeta.R
\name{getConfiguredEntityTypes}
\alias{getConfiguredEntityTypes}
\title{Gets configured entity type}
\usage{
getConfiguredEntityTypes(displayName)
}
\arguments{
\item{displayName}{character Optional display name to filter}
}
\value{
list Configured entity types
}
\description{
Get configured entity type
}
\details{
Gets and optionally filters by display name the list of configured entity types in acas
}
\keyword{entity,}
\keyword{meta}

