% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JSON_library.R
\name{getDDictKinds}
\alias{getDDictKinds}
\title{getDdictKinds}
\usage{

  getDDictKinds(lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{lsServerURL}{(racas::applicationSettings$client.service.persistence.fullpath)}
}
\value{
a data frame of kinds
}
\description{
Get ddict kinds as described by URLencode(paste0(racas::applicationSettings$client.service.persistence.fullpath,"/api/v1/ddictkinds"))
}

