% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JSON_library.R
\name{getDDictValuesByTypeKindFormat}
\alias{getDDictValuesByTypeKindFormat}
\title{Get ddict values by type and kind}
\usage{
getDDictValuesByTypeKindFormat(lsKind, lsType, format = "json",
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{lsKind}{}

\item{lsType}{}

\item{format}{(json, tsv)}
}
\value{
A data.frame or json of ddict values
}
\description{
Gets all ddict values by type and kind
}

