% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getEntity}
\alias{getEntity}
\title{Retrieves a list of entities (experiments or prototocols) by type}
\usage{
getEntity(x, type = c("protocolName", "experimentName", "protocolCodeName",
  "experimentCodeName"))
}
\arguments{
\item{x}{entity names or code names of the same type}

\item{type}{entity type of x (protocolName, protocolCodeName, experimentName or experimentCodeName)}
}
\value{
a list of the entities returned from the server
}
\description{
This function retrieves a list of protocols or experiments by code name or label
}
\examples{
getEntity("ADME_Human Liver Microsome Stability", "protocolName")
getEntity(c("ADME_Human Liver Microsome Stability","some other protocol name"), "protocolName")
getEntity("PROT-00000001", "protocolCodeName")
getEntity("EXPT-00000001", "experimentCodeName")
}
\keyword{entity}
\keyword{get,}

