% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getEntityById}
\alias{getEntityById}
\title{Gets an entity by id}
\usage{
getEntityById(id, entityKind, include = "", errorEnv = NULL,
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{entity}{an entity object, such as an experiment or analysis group}

\item{entityKind}{the kind of entity, such as "experiment" or
"analysisgroup", see \link{acasEntityHierarchy}}

\item{include}{a character string describing what to include}

\item{lsServerURL}{the url for the roo server}
}
\value{
a named list representing an object.
}
\description{
Gets an entity object by id and kind
}
\details{
This will fail with an error if the object does not exist.
  \code{include} is only implemented for some entities, see
  \link{getExperimentById} for more detail.
}

