% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getErrorText}
\alias{getErrorText}
\title{Get error text for users}
\usage{
getErrorText(errorList)
}
\arguments{
\item{errorList}{a list of items of class "error"}
}
\description{
From a list of error objects, get error messages for users, replacing
internal errors with a message to look in the logs.
}
\details{
Error messages with class \code{\link{userStop}} pass just their
  message, future plans to group errors by class if there are very large
  numbers.
}

