% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JSON_library.R
\name{getExperimentStatesByTypeAndKind}
\alias{getExperimentStatesByTypeAndKind}
\title{Get Experiment States}
\usage{
getExperimentStatesByTypeAndKind(experimentId, stateType, stateKind,
  responseFormat = "json",
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{experimentId}{id or codeName of the experiment}

\item{stateType}{lsType of the state}

\item{stateKind}{lsKind of the state}

\item{responseFormat}{"json" or maybe "tsv"}
}
\description{
Get states by type and kind.
}
\details{
Returns a list of experiment states.
}

