% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JSON_library.R
\name{getExperimentValuesByTypeAndKind}
\alias{getExperimentValuesByTypeAndKind}
\title{Get Experiment Values}
\usage{
getExperimentValuesByTypeAndKind(experimentId, stateType, stateKind, valueType,
  valueKind, responseFormat = "json",
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{experimentId}{id or codeName of the experiment}

\item{stateType}{lsType of the state}

\item{stateKind}{lsKind of the state}

\item{valueType}{lsType of the value}

\item{valueKind}{lsKind of the value}

\item{responseFormat}{"json" or maybe "tsv"}
}
\value{
List of experiment values.
}
\description{
Get values without requiring knowledge of whether the value already
exists or not- it will be checked by the roo server.
}

