% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFileEncoding}
\alias{getFileEncoding}
\title{Check file encoding}
\usage{
getFileEncoding(filePath)
}
\arguments{
\item{filePath}{The file that needs to be checked, relative to your working directory}
}
\value{
The file encoding for the input file

Example UTF-16LE file found in inst/test/data/UTF-16LE-test
}
\description{
Checks to see the encoding of the input file.
}

