% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JSON_library.R
\name{getOrCreateDDictKinds}
\alias{getOrCreateDDictKinds}
\title{getOrCreateDDictKinds}
\usage{
getOrCreateDDictKinds(typesKindsDataFrame,
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{typesKindsDataFrame}{(described here RCurl::curlEscape(paste0(racas::applicationSettings$client.service.persistence.fullpath,"/api/v1/setup/ddictkinds")))}
}
\value{
list of types and kinds
}
\description{
Registers ddict kinds from json
}

