% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JSON_library.R
\name{getOrCreateEntityState}
\alias{getOrCreateEntityState}
\title{Gets a state}
\usage{
getOrCreateEntityState(entity, entityKind, stateType, stateKind, recordedBy,
  lsTransaction,
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{entity}{an entity object, such as an experiment or analysis group}

\item{entityKind}{the kind of entity, such as "experiment" or "analysisgroup", see \link{acasEntityHierarchy}}

\item{stateType}{lsType of the state}

\item{stateKind}{lsKind of the state}

\item{recordedBy}{the current username}

\item{lsTransaction}{the id of the transaction}

\item{lsServerURL}{the url for the roo server}
}
\description{
Gets a state by stateKind if it exists, or creates a new one if not
}
\details{
This will fail if the entity has more than one non-ignored state of entered stateKind, as it would be unclear which to update.
}

