% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getOrCreateExperimentState}
\alias{getOrCreateExperimentState}
\title{Gets an experiment state}
\usage{
getOrCreateExperimentState(experiment, stateType, stateKind, recordedBy,
  lsTransaction,
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{experiment}{an experiment object}

\item{stateType}{lsType of the state}

\item{stateKind}{lsKind of the state}

\item{recordedBy}{the current username}

\item{lsTransaction}{the id of the transaction}

\item{lsServerURL}{the url for the roo server}
}
\description{
Gets an experiment state by stateKind if it exists, or creates a new one if
not.
}
\details{
This will fail if the experiment has more than one non-ignored state
  of entered stateKind, as it would be unclear which to update.
}

