% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preferredBatchCode.R
\name{getPreferredId2}
\alias{getPreferredId2}
\title{Gets preferred ids}
\usage{
getPreferredId2(entityIds, entityType, entityKind, preferredIdService = NULL)
}
\arguments{
\item{entityType}{the type of the id's, e.g. "compound"}

\item{entityKind}{the kind of the id's, e.g. "batch name"}

\item{preferredIdService}{the url of the preferred id service. Defaults to
\code{paste0(racas::applicationSettings$server.nodeapi.path,
"/api/entitymeta/preferredCodes")}}

\item{batchIds}{a character vector of batch codes}
}
\value{
a data.frame with names \code{c("Requested.Name", "Preferred.Code")}
}
\description{
Entered entity codes (ids) are checked against a preferred id service.
}
\details{
Gets preferred id's for a range for inputs. For compound/batch name
  requests, it will use the relevant batch code check. For lsThings, it
  checks those entities for their preferred codes.
}
\keyword{batchCode,}
\keyword{preferred}

