% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getPreferredIdInternal}
\alias{getPreferredIdInternal}
\title{Gets preferred ids}
\usage{
getPreferredIdInternal(batchIds, preferredIdService = NULL,
  testMode = FALSE)
}
\arguments{
  \item{batchIds}{a character vector of batch codes}

  \item{preferredIdService}{the url of the preferred id
  service. Defaults to
  \code{racas::applicationSettings$preferredBatchIdService}}

  \item{testMode}{deprecated, was used for a test mode}
}
\value{
a list of pairs of requested IDs and preferred IDs. On
error, empty list
}
\description{
Internal code for \link{getPreferredId} (before chunking)
}

