% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JSON_library.R
\name{getProtocolsByName}
\alias{getProtocolsByName}
\title{Protocol search by name}
\usage{
getProtocolsByName(protocolName,
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{protocolName}{a string, the name of the protocol}

\item{lsServerURL}{url for roo server}
}
\value{
a list of protocols
}
\description{
Gets protocols by name
}
\details{
returns a list as uniqueness is not always enforced
}

