% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getURLcheckStatus}
\alias{getURLcheckStatus}
\title{Get URL and check status}
\usage{
getURLcheckStatus(url, ...)
}
\arguments{
\item{url}{the url to get/post}

\item{...}{optional parameters passed to getURL}
}
\description{
This is a wrapper for getURL that throws an error when the HTTP status is 400 or greater
}
\details{
checks the HTTP status and logs to racas.log as com.acas.sel if 400 or greater. See also \link{postURLcheckStatus}.
}

