\name{getValuesByTypeAndKind}
\alias{getValuesByTypeAndKind}
\title{Gets values from a state}
\usage{
  getValuesByTypeAndKind(lsState, typeAndKind)
}
\arguments{
  \item{lsState}{A list that represents a single lsState}

  \item{typeAndKind}{A string value that follows the form
  "lsType_lsKind"}
}
\value{
  A list of lsValues
}
\description{
  Allows you to select states of a specific lsTypeAndKind
  from a state
}

