% Generated by roxygen2 (4.0.0): do not edit by hand
\name{interpretJSONBoolean}
\alias{interpretJSONBoolean}
\title{Turns 'true' and 'false' into TRUE and FALSE}
\usage{
interpretJSONBoolean(JSONBoolean)
}
\arguments{
\item{JSONBoolean}{a string of "true" or "false"

Other inputs not affected}
}
\description{
Turns 'true' and 'false' into TRUE and FALSE
}

