% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/longFormatSave.R
\name{meltConcentrations2}
\alias{meltConcentrations2}
\title{Turns concentration columns into rows in a long format}
\usage{
meltConcentrations2(entityData)
}
\arguments{
\item{entityData}{a data frame with data, must include rows
"concentration", "concentrationUnit", "tempStateId",
"parentId", "tempId", "stateType", "stateKind"

For use with Tsv saves}
}
\description{
Turns concentration columns into rows in a long format
}

