% Generated by roxygen2 (4.0.0): do not edit by hand
\name{meltTimes}
\alias{meltTimes}
\title{Turns time columns into rows in a long format}
\usage{
meltTimes(entityData, entityKind = "treatmentGroup")
}
\arguments{
\item{entityData}{a data frame with data, must include rows "time" and "timeUnit"}
}
\description{
Turns time columns into rows in a long format
}

