% Generated by roxygen2 (4.0.1): do not edit by hand
\name{meltTimes2}
\alias{meltTimes2}
\title{Turns time columns into rows in a long format}
\usage{
meltTimes2(entityData)
}
\arguments{
\item{entityData}{a data frame with data, must include rows
"time", "timeUnit", "tempStateId",
"parentId", "tempId", "stateType", "stateKind"

For use with Tsv saves}
}
\description{
Turns time columns into rows in a long format
}

