% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_fit.R
\docType{class}
\name{ModelFit-class}
\alias{ModelFit}
\alias{ModelFit-class}
\title{ModelFit}
\description{
ModelFit
}
\section{Fields}{

\describe{
\item{\code{drc_function}}{function.}

\item{\code{paramNames}}{character.}

\item{\code{categorization_function}}{function.}

\item{\code{get_reported_parameters}}{function.}

\item{\code{apply_limits}}{function.}

\item{\code{default_fit_settings}}{list.}

\item{\code{simple_to_advanced_fittings_function}}{function.}

\item{\code{model_equation_img}}{character.}
}}

