% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JSON_library.R
\name{modules}
\alias{modules}
\title{Known acas modules which require lsKinds to exist}
\usage{
modules()
}
\value{
a list of modules which require lsKinds to be registered
}
\description{
There are a set of required lsKinds that need to be loaded in modules. This function returns the module names for which there are required lsKinds that need to be registered
See \link{load_value_type_and_kinds} to load lsKinds
}

