% Generated by roxygen2 (4.0.2): do not edit by hand
\name{moveFileToFileServer}
\alias{moveFileToFileServer}
\title{Move file to file server}
\usage{
moveFileToFileServer(fileStartLocation, targetPath = NULL,
  fileServiceType = racas::applicationSettings$server.service.external.file.type,
  fileService = racas::applicationSettings$server.service.external.file.service.url,
  experiment = NULL, recordedBy = NULL)
}
\arguments{
\item{fileStartLocation}{current location of file, relative from privateUploads}

\item{targetPath}{path to new file location, not currently used when fileServiceType == "custom"}

\item{fileServiceType}{"blueimp" or "custom"}

\item{fileService}{path to file service, not used when fileServiceType == "blueimp"}

\item{experiment}{experiment object, not used when fileServiceType == "blueimp"}

\item{recordedBy}{logged in username, not used when fileServiceType == "blueimp"}
}
\value{
character file code or path relative from privateUploads
}
\description{
Move file to file server
}
\keyword{blueimp}
\keyword{custom,}
\keyword{file,}

