% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tsvSave.R
\name{naIfNull}
\alias{naIfNull}
\title{Turn NULL into NA}
\usage{
naIfNull(x, naType = NULL)
}
\arguments{
\item{x}{input}

\item{naType}{NA value, e.g.\code{NA_character_}}
}
\description{
Useful for changing objects to data frames
}
\details{
will return the input for any value other than NULL
}

