% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JSON_library.R
\name{pickBestLabel}
\alias{pickBestLabel}
\title{Get preferred labels}
\usage{
pickBestLabel(entity, labelTypeAndKind = NA)
}
\arguments{
\item{entity}{a list entity, such as a protocol or experiment.}

\item{labelTypeAndKind}{a labelTypeAndKind such as "name_protocol name".}
}
\value{
A label list object.
}
\description{
Gets the best label, possibly limited to a labelTypeAndKind. "Best" is
defined as the preferred label, or if that fails, the most recent one. This
is translated from coffeescript Label.coffee.
}

