% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JSON_library.R
\name{pickBestName}
\alias{pickBestName}
\title{Pick Best Name}
\usage{
pickBestName(entity)
}
\arguments{
\item{entity}{a list that is a protocol, experiment, etc.}
}
\value{
a list that is a label
}
\description{
From an acas entity (protocol, experiment, etc.), get the label that is the
best name
}

