% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query.R
\name{query_replace_string_with_values}
\alias{query_replace_string_with_values}
\title{Replace character in query string with values over n times ceiling(values/limit) and return query results for each resulting query}
\usage{
query_replace_string_with_values(qu, string, values, limit = 999, ...)
}
\arguments{
\item{qu}{a sql query character string that contains the character given in string parameter}

\item{string}{a character string to replace with values parameter}

\item{values}{a list of values to replace in string parameter}

\item{...}{expressions fed to \code{\link{query}}}
}
\value{
A list objects with results from query
}
\description{
Function originally created to get around oracle in clause query (1000).  Calls \code{\link{query}} after replacing a specified character with a subset of values
}
\examples{
query_replace_string_with_values(qu = "select * from analysis_group_value where id in (REPLACEME)", string = "REPLACEME", values = c(1:1001))
}
\keyword{inclause,}
\keyword{oracle,}
\keyword{query,}
\keyword{workaround}

