% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/updateExperimentMetadata.R
\name{saveAcasFile}
\alias{saveAcasFile}
\title{save file in ACAS}
\usage{
saveAcasFile(fileStartLocation, entity, entityKind, stateType, stateKind,
  valueKind, recordedBy, lsTransaction, valueType = "fileValue",
  additionalPath = "", experimentCodeName = NULL,
  fileServiceType = racas::applicationSettings$server.service.external.file.type,
  fileService = racas::applicationSettings$server.service.external.file.service.url,
  deleteOldFile = TRUE, customSourceFileMove = NULL,
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{fileStartLocation}{path to file, relative from privateUploads}

\item{entity}{an entity object, such as an experiment}

\item{entityKind}{the kind of entity, such as "experiment" or
"analysisgroup", see \link{acasEntityHierarchy}}

\item{stateType}{the stateType of save location, e.g. "metadata"}

\item{stateKind}{the stateKind of save location, e.g. "experiment metadata"}

\item{valueKind}{the valueKind, or description of the file e.g. "source file"}

\item{recordedBy}{username of current user}

\item{lsTransaction}{the id of the transaction}

\item{valueType}{usually fileValue, could also be inlineFileValue}

\item{additionalPath}{if saving to a file system and not a service,
additional file organization, i.e. "analysis/final"}

\item{fileServiceType}{"blueimp" for internal, "custom" for custom}

\item{fileService}{only required if fileServiceType is "custom", path to
custom file service}

\item{deleteOldFile}{boolean, \code{TRUE} if current file should be deleted
after a copy is made}

\item{customSourceFileMove}{function for custom moving file to server}

\item{lsServerURL}{the url for the roo server}
}
\description{
Saves a file into a file service and saves a reference to it in the provided
entity
}
\details{
sets the "status" value in state "analysis status" (or "dryrun
  status" if \code{dryRun == TRUE}). Mostly used by
  \code{\link{saveAcasFileToExperiment}}. In updateExperimentMetadata.R
}

