% Generated by roxygen2 (4.0.2): do not edit by hand
\name{saveContainer}
\alias{saveContainer}
\title{save container objects
Currently, this cannot accept labels and states}
\usage{
saveContainer(container,
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\description{
save container objects
Currently, this cannot accept labels and states
}

