% Generated by roxygen2 (4.0.2): do not edit by hand
\name{saveDataDirectDatabase}
\alias{saveDataDirectDatabase}
\title{Save data tables direct to database}
\usage{
saveDataDirectDatabase(agData, tgData, subjectData, lsTransactionId = NA,
  experimentId = NULL)
}
\arguments{
\item{agData}{data.frame or data.table of analysis group data}

\item{tgData}{data.frame or data.table of treatment group data}

\item{subjectData}{data.frame or data.table of subject data}

\item{lsTransactionId}{integer id of transaction}

\item{experimentId}{integer id of experiment}
}
\description{
Saves whole sets of data for an experiment to the database, and rolls back
the transaction if anything fails (or deletes on Oracle).
}
\details{
This does not work through any roo service, so it will be faster,
  but has fewer integrity checks along the way. Use with caution. For
  required columns, see example with \code{data(agData)} or look in code of
  \code{\link{prepareTableForDD}}. If you need to do updates as well as
  saves, use \code{\link{saveAllViaTsv}}. The experimentId and
  lsTransactionId will fill in if they are missing in the input data tables.
}

