% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/updateExperimentMetadata.R
\name{saveInputParameters}
\alias{saveInputParameters}
\title{Save experiment parameters}
\usage{
saveInputParameters(inputParameters, experiment, lsTransaction, recordedBy)
}
\arguments{
\item{inputParameters}{string that contains JSON}

\item{experiment}{list, experiment object from JSON, only needs an id}

\item{lsTransaction}{unused}

\item{recordedBy}{unused}
}
\description{
Save the experiment parameters passed from the GUI to the data analysis parameters clobValue.
}

