% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/longFormatSave.R
\name{saveLabelsFromLongFormat}
\alias{saveLabelsFromLongFormat}
\title{Save labels from a long format}
\usage{
saveLabelsFromLongFormat(entityData, entityKind, stateGroups, idColumn,
  recordedBy, lsTransaction, stateGroupIndices = NULL, labelPrefix = NULL)
}
\arguments{
\item{entityData}{a data.frame, including one column named 'stateGroupIndex', one that matches idColumn, and one of the form 'entityID'}

\item{entityKind}{a string, the kind of the state, limited to "protocol", "experiment", "analysisgroup", "treatmentgroup", "subject", "container", "itxcontainercontainer"}

\item{stateGroups}{a list of lists, each of which includes details about how to save states (TODO link later)}

\item{idColumn}{a string, the name of the column used to separate states (often stateID)}

\item{recordedBy}{a string, the name of the person recording the data}

\item{lsTransaction}{an integer, the id of the lsTransaction}

\item{stateGroupIndices}{an integer vector of indices to use from stateGroups}

\item{labelPrefix}{a string, prefixed to all labels}
}
\description{
This function saves labels to the database specified in \code{\link{applicationSettings}}
}
\keyword{format,}
\keyword{label,}
\keyword{labels}
\keyword{save,}
\keyword{stateGroups,}

