% Generated by roxygen2 (4.0.2): do not edit by hand
\name{saveTsvData}
\alias{saveTsvData}
\title{Save from tsv}
\usage{
saveTsvData(experimentId, lsTransactionId, agDatafile, tgDataFile,
  subjectDataFile)
}
\description{
This function was used for a migration from flat files, and could be used for
others later. This would then be reworked to use
\code{\link{saveDataDirectDatabase}} and \code{\link{readTsvDataFileDD}}.
}

