% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/directDatabaseSave.R
\name{saveValuesDD}
\alias{saveValuesDD}
\title{Save values direct database}
\usage{
saveValuesDD(conn, entityType, inputDT)
}
\description{
This function was used for a migration from flat files, and could be used for
others later. This would then be reworked to use
\code{\link{saveDataDirectDatabase}} and \code{\link{readTsvDataFileDD}}.
}

