% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/updateExperimentMetadata.R
\name{setExperimentHtml}
\alias{setExperimentHtml}
\title{Experiment html}
\usage{
setExperimentHtml(htmlText, experiment, recordedBy, dryRun = F,
  lsTransaction = NULL,
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{htmlText}{html element to display for experiment}

\item{experiment}{an experiment object}

\item{recordedBy}{username of current user}

\item{dryRun}{boolean if in dry run}

\item{lsTransaction}{the id of the transaction}

\item{lsServerURL}{the url for the roo server}
}
\description{
Sets experiment html
}
\details{
sets the "status" value in state "analysis status" (or "dryrun
  status" if \code{dryRun == TRUE}). In updateExperimentMetadata.R
}

