% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setExperimentStatus}
\alias{setExperimentStatus}
\title{Experiment status}
\usage{
setExperimentStatus(status, experiment, recordedBy, dryRun = F,
  lsTransaction = NULL,
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{status}{a string such as "running" or "completed"}

\item{experiment}{an experiment object}

\item{recordedBy}{username of current user}

\item{dryRun}{boolean if in dry run}

\item{lsTransaction}{the id of the transaction}

\item{lsServerURL}{the url for the roo server}
}
\description{
Sets experiment status
}
\details{
sets the "status" value in state "analysis status" (or "dryrun
  status" if \code{dryRun == TRUE}). In updateExperimentMetadata.R
}

