% Generated by roxygen2 (4.0.0): do not edit by hand
\name{simpleToAdvancedFitSettings}
\alias{simpleToAdvancedFitSettings}
\title{Convert a simple dose response request to an advanced dose response request}
\usage{
simpleToAdvancedFitSettings(simpleSettings, updateFlags = NULL,
  modelHint = "LL.4")
}
\arguments{
\item{simpleSettings}{a list object of simple fit settings}

\item{modelHint}{a character string to identify the dose response model to fit}
}
\value{
an advanced fit settings list object
}
\description{
Reads the default fit settings for the given model hint and updates it based on the simple request
}
\examples{
file <- system.file("docs", "example-simple-fitsettings-ll4.json", package = "racas")
simpleSettingsJSON <- readChar(file, file.info(file)$size)
simpleSettings <- fromJSON(simpleSettingsJSON)
simpleToAdvancedFitSettings(simpleSettings)
}

