% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stopUser}
\alias{stopUser}
\title{Fatal error tracking}
\usage{
stopUser(message)
}
\arguments{
  \item{message}{The error message that the user should
  see}
}
\value{
Stops the function, and adds the class "userStop" to the
error object

All helpful errors should be thrown using \code{stopUser}.
Any error thrown using \code{stop} will be treated as an
internal error by the simple experiment loader
}
\description{
Adds the class "userStop" to a fatal error, to
differentiate between errors that should be displayed to
users and errors that shouldn't (internal errors)
}

