% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/errorLogging.R
\name{tryCatch.W.E}
\alias{tryCatch.W.E}
\title{Error and warning trapping}
\usage{
tryCatch.W.E(expr)
}
\arguments{
\item{expr}{An expression which will have errors caught}
}
\value{
A list with value (either result or error) and warningList
}
\description{
Traps errors and warnings, creating lists of each
}

