% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tsv_url_to_data_table}
\alias{tsv_url_to_data_table}
\title{tsv service url to data.table}
\usage{
tsv_url_to_data_table(url, type = c("simple", "complex"), ...)
}
\arguments{
\item{url}{a url encoded string that calls a service that returns a tsv}

\item{type}{simple (uses data.table's fread which is fast but can't handle embedded html or \t values) or complex (uses read.csv which is slower but can handle embeded tables .etc.)}
}
\value{
a data.table result
}
\description{
Calls a tsv service and returns a data.table of results
}
\examples{
# AG values for dose response have complex results
url <- "http://host4.labsynch.com:8080/acas/api/v1/experiments/EXPT-00000408/agvalues/bystate/data/Dose\%20Response/tsv"
tsv_url_to_data_table(url, "complex")

Subject values fo rdose response have simple results
url <- "http://host4.labsynch.com:8080/acas/api/v1/experiments/EXPT-00000408/subjectvalues/bystate/data/results/tsv"
tsv_url_to_data_table(url, "simple")
}

