% Generated by roxygen2 (4.0.0): do not edit by hand
\name{updateAcasEntity}
\alias{updateAcasEntity}
\title{Updates an entity}
\usage{
updateAcasEntity(entity, acasCategory,
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{entity}{the entity to place}

\item{acasCategory}{the category (e.g. "experiments", "containervalues")}

\item{lsServerURL}{the URL of the persistence server}
}
\description{
Replaces the entity that is at the URL with the one sent. Sub-entities (label, state, value) must have a parent object
}

