% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JSON_library.R
\name{updateValueByTypeAndKind}
\alias{updateValueByTypeAndKind}
\title{Update Values}
\usage{
updateValueByTypeAndKind(newValue, entityKind, parentId, stateType, stateKind,
  valueType, valueKind,
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{newValue}{value to save, will be sent as a string}

\item{entityKind}{kind of entity, e.g. "experiment"}

\item{parentId}{id of the parent entity}

\item{stateType}{lsType of the state}

\item{stateKind}{lsKind of the state}

\item{valueType}{lsType of the value}

\item{valueKind}{lsKind of the value}
}
\value{
updated value object
}
\description{
Updates values without requiring knowledge of whether the value already
exists or not- it will be checked by the roo server. Also adds valueType and
valueKind if needed.
}

