% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JSON_library.R
\name{validateValueKindsFromDataFrame}
\alias{validateValueKindsFromDataFrame}
\title{validateValueKindsFromDataFrame}
\usage{
validateValueKindsFromDataFrame(typesAndKindsDataFrame,
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{typesAndKindsDataFrame}{data.frame with 2 columns: \code{lsType} and \code{lsKind}}

\item{lsServerURL}{server to check against, currently ignored}
}
\value{
a data.table with columns \code{lsTypeName} (character), \code{lsKindName} (character), \code{lsKind} (full object kinds), and \code{lsKindExists} (boolean).
}
\description{
Get a data.table of type names and kind names plus the kind (full object) from a data.frame of type names and kind names.
}
\details{
Column \code{lsKind} will be NULL if there is no value kind found.
}
\examples{
# Not run because this needs a server, example output is below.
# output <- validateValueKindsFromDataFrame(data.frame(lsType="numericValue", lsKind="time"))
output <- structure(list(
 lsTypeName = "numericValue", lsKindName = "time",
 lsKind = list(list(structure(list(id = 12, kindName = "time",
                                   lsType = structure(list(id = 7, typeName = "numericValue", version = 0),
                                                      .Names = c("id", "typeName", "version")),
                                   lsTypeAndKind = "numericValue_time", version = 0),
                              .Names = c("id", "kindName", "lsType", "lsTypeAndKind", "version")))),
 lsKindExists = TRUE), .Names = c("lsTypeName", "lsKindName", "lsKind", "lsKindExists"),
 sorted = c("lsTypeName", "lsKindName"), class = c("data.table", "data.frame"),
 row.names = c(NA, -1L))
}

