% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{acasEntityHierarchy}
\alias{acasEntityHierarchy}
\alias{acasEntityHierarchyCamel}
\title{The hierarcy of ACAS entities}
\format{\preformatted{ chr [1:5] "protocol" "experiment" "analysisgroup" "treatmentgroup" ...
}}
\usage{
acasEntityHierarchy

acasEntityHierarchyCamel
}
\description{
Each category of entity contains the level below. Both a
lowercase version (for URLs) and a camelCase verison (for
JSON) exist. Can be found in racas::acasEntityHierarchy and
racas::acasEntityHierarchyCamel
}
\keyword{datasets}

