% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addError}
\alias{addError}
\alias{userError}
\title{Non-fatal error tracking}
\usage{
addError(errorMessage, errorEnv = NULL)
}
\arguments{
\item{errorMessage}{The message for the user
param errorEnv deprecated

The error is added as class \code{userError} to the racasMessenger.}
}
\description{
Add errors to a list of errors to show a user
}

