% Generated by roxygen2 (4.0.1): do not edit by hand
\name{api_createProtocol}
\alias{api_createProtocol}
\title{Create acas protocol}
\usage{
api_createProtocol(protocolName, shortDescription, recordedBy,
  stopOnAnyExist = TRUE)
}
\arguments{
\item{protocolName}{}

\item{shortDescription}{}
}
\value{
recordedBy
}
\description{
This function creates an acas protocol
}
\examples{
api_createProtocol(c("Brian Test Protocol2"), shortDescription = "Test Protocol", recordedBy = "bbolt")
}
\keyword{create,}
\keyword{protocol}
\keyword{register,}

