% Generated by roxygen2 (4.0.0): do not edit by hand
\name{changeEntityMode}
\alias{changeEntityMode}
\title{Change mode ACAS entity}
\usage{
changeEntityMode(entityKind, currentMode, desiredMode)
}
\arguments{
\item{entityKind}{A list, one of racas::acasEntityHierarchy (or
acasEntityHierarchyCamel or acasEntityHierarchySpace)}

\item{currentMode}{One of "lowercase", "camel", or "space"}

\item{desiredMode}{One of "lowercase", "camel", or "space"}
}
\description{
Changes the mode of an acas entity from camelcase to all lowercase or with a
space between words
}

