% Generated by roxygen2 (4.0.0): do not edit by hand
\name{createHtmlSummary}
\alias{createHtmlSummary}
\title{Create an HTML summary}
\usage{
createHtmlSummary(hasError, errorList, hasWarning, warningList, summaryInfo,
  dryRun)
}
\arguments{
  \item{hasError}{A boolean marking if the function
  returned an error}

  \item{errorList}{A list of errors}

  \item{hasWarning}{A boolean marking if the function had
  warnings}

  \item{warningList}{A list of warnings}

  \item{summaryInfo}{A list of information to return to the
  user (see details)}

  \item{dryRun}{A boolean marking if information should be
  saved to the server}
}
\value{
a string of HTML
}
\description{
Takes a list of summary information and turns it into HTML
that can be displayed in a web interface.
}
\details{
summaryInfo can have named elements "info": a list of
information where names are used as titles, and
"viewerLink": a URL link to the program used to view the
results
}

