% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{createSELRawResults}
\alias{createSELRawResults}
\title{Creates a data.frame that contains the elements of the Raw Results section of SEL}
\usage{
createSELRawResults(rawResults, curveIDColumn = "curve id", ls_kinds = NULL,
  unit_kinds = NULL)
}
\description{
The return structure represents the standard inputs of the
"Raw Results" section of SEL
}
\examples{
Not implemented yet
}

