% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createStateValue}
\alias{createStateValue}
\title{Creates a state value}
\usage{
createStateValue(lsType = "lsType", lsKind = "lsKind", stringValue = NULL,
  fileValue = NULL, urlValue = NULL, publicData = TRUE, ignored = FALSE,
  dateValue = NULL, clobValue = NULL, blobValue = NULL,
  valueOperator = NULL, operatorType = NULL, numericValue = NULL,
  sigFigs = NULL, uncertainty = NULL, uncertaintyType = NULL,
  numberOfReplicates = NULL, valueUnit = NULL, unitType = NULL,
  comments = NULL, lsTransaction = NULL, codeValue = NULL,
  recordedBy = "username", lsState = NULL, testMode = FALSE,
  recordedDate = as.numeric(format(Sys.time(), "\%s")) * 1000,
  codeType = NULL, codeKind = NULL, codeOrigin = NULL)
}
\arguments{
\item{lsType}{type of the value}

\item{lsKind}{lsKind of the value}

\item{stringValue}{<255 character}

\item{fileValue}{file code or path}

\item{urlValue}{url}

\item{publicData}{TRUE to be visible}

\item{ignored}{TRUE to mark as old}

\item{dateValue}{date in milliseconds}

\item{clobValue}{clob}

\item{blobValue}{blob}

\item{valueOperator}{">" or "<"}

\item{operatorType}{"comparison", not yet implemented}

\item{numericValue}{numeric}

\item{sigFigs}{integer}

\item{uncertainty}{numeric}

\item{uncertaintyType}{"standard deviation"}

\item{numberOfReplicates}{integer}

\item{valueUnit}{"uM", etc.}

\item{unitType}{not yet implemented}

\item{comments}{used by fileValue for a filename, flags for comments, etc.}

\item{lsTransaction}{id of the transaction}

\item{codeValue}{codename of something}

\item{recordedBy}{the current username}

\item{lsState}{a state object}

\item{testMode}{used for testing}

\item{testMode}{used for testing}

\item{lsServerURL}{the url for the roo server}
}
\value{
list, a value object
}
\description{
Creates a state value, can include an lsState or be nested inside one.
}
\details{
Use either in a nested object or alone
}

