% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JSON_library.R
\name{createSubjectContainerInteraction}
\alias{createSubjectContainerInteraction}
\title{create subject container interaction}
\usage{
createSubjectContainerInteraction(subject, container, lsType,
  lsKind = "interaction", codeName = NULL, ignored = FALSE,
  lsTransaction = NULL, recordedBy = "userName", interactionStates = NULL)
}
\description{
create subject container interaction
}
\details{
interactionStates is ignored for now,
could add back later (as lsStates) with roo update.

This is super sensitive to subject and container being nested- only known to
work with containers and subjects that only have an id and version.
}

