% Generated by roxygen2 (4.0.0): do not edit by hand
\name{createTag}
\alias{createTag}
\title{Creates a tag}
\usage{
createTag(tagText, id = NULL, version = NULL)
}
\arguments{
\item{tagText}{the text of the tag}

\item{id}{used to link to old tags}

\item{version}{the version of the tag (only used with an id)}
}
\description{
Creates a tag
}

