% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JSON_library.R
\name{deleteAnalysisGroupsByExperiment}
\alias{deleteAnalysisGroupsByExperiment}
\title{Delete analysis groups by experiment}
\usage{
deleteAnalysisGroupsByExperiment(experiment,
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{experiment}{a list that has an element id for the experiment}

\item{lsServerURL}{the URL of the persistence server}
}
\value{
empty string
}
\description{
Deletes all analysis groups within an experiment
}
\details{
Deletes all of the analysis groups, even if they are linked to other
  experiments. This is intended for fully clearing uploaded data. Does not
  mark treatment groups and subjects as deleted, retrieval functions should
  respect parent deletion.
}

