% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_or_create_value_kinds}
\alias{get_or_create_value_kinds}
\title{Get Or Create Value Kind}
\usage{
get_or_create_value_kinds(df,
  persistence_full_path = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{a}{data frame (or data table) with columns lsType and lsKind}
}
\value{
a list object of returned lsKinds
}
\description{
Gets or creates a set of value kinds given a data frame of lsType (name only) and lsKind (name)
}

