% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_text_file_contents}
\alias{get_text_file_contents}
\title{Get text file contents as character}
\usage{
get_text_file_contents(file_path)
}
\arguments{
\item{file}{character path to file}
}
\value{
character contents of file
}
\description{
Get text file contents as character
}
\keyword{character,}
\keyword{file}
\keyword{read,}
\keyword{text,}

