% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getAllValueKinds}
\alias{getAllValueKinds}
\title{valueKinds}
\usage{

  getAllValueKinds(lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{lsServerURL}{url for roo server}
}
\description{
Gets a list of all valueKinds available.
}

