% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getAllValueTypes}
\alias{getAllValueTypes}
\title{valueTypes}
\usage{

  getAllValueTypes(lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{lsServerURL}{url for roo server}
}
\description{
Gets a list of all valueTypes available.
}

