% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getExperimentsByName}
\alias{getExperimentsByName}
\title{Experiment search by name}
\usage{
getExperimentsByName(experimentName,
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{experimentName}{a string, the name of the experiment}

\item{lsServerURL}{url for roo server}
}
\value{
a list of experiments
}
\description{
Gets experiments by name
}
\details{
returns a list as uniqueness is not always enforced
}

