% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFileEncoding}
\alias{getFileEncoding}
\title{Check file encoding}
\usage{
getFileEncoding(filePath)
}
\arguments{
\item{filePath}{The file that needs to be checked, relative to your working directory}
}
\value{
The file encoding for the input file

Two example UTF-16LE file can be found in inst/tests/data/:

UTF-16LE-test.txt

UTF-16LE-test2.txt
}
\description{
Checks to see the encoding of the input file.
}

