% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JSON_library.R
\name{getOrCreateDDictTypes}
\alias{getOrCreateDDictTypes}
\title{getOrCreateDDictTypes}
\usage{
getOrCreateDDictTypes(typesList,
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{list}{(described here RCurl::curlEscape(paste0(racas::applicationSettings$client.service.persistence.fullpath,"/api/v1/setup/ddicttypes")))}
}
\value{
list of types
}
\description{
Registers ddict types from json
}

