% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preferredBatchCode.R
\name{getPreferredId}
\alias{getPreferredId}
\title{Gets preferred ids}
\usage{
getPreferredId(batchIds, preferredIdService = NULL, testMode = FALSE)
}
\arguments{
\item{batchIds}{a character vector of batch codes}

\item{preferredIdService}{the url of the preferred id service. Defaults to \code{racas::applicationSettings$preferredBatchIdService}}

\item{testMode}{deprecated, was used for a test mode}
}
\value{
a list of pairs of requested IDs and preferred IDs. On error, empty list
}
\description{
Entered batch codes (ids) are checked against a preferred id service
}
\details{
DEPRECATED: use \code{\link{getPreferredId2}} with \code{entityType="compound"} and \code{entityKind="batch name"}.
}
\keyword{batchCode,}
\keyword{preferred}

