% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getPreferredLabelText}
\alias{getPreferredLabelText}
\title{Get preferred label text}
\usage{
getPreferredLabelText(entity, labelTypeAndKind = NA)
}
\arguments{
\item{entity}{a list entity, such as a protocol or experiment.}

\item{labelTypeAndKind}{a labelTypeAndKind such as "name_protocol name".}
}
\value{
A text string.
}
\description{
Gets the preferred name of an entity.
}

