\name{getUploadedFilePath}
\alias{getUploadedFilePath}
\title{Returns file path for uploaded files}
\usage{
getUploadedFilePath(inputString)
}
\arguments{
  \item{inputString}{File Name}
}
\value{
File path
}
\description{
Returns file path for uploaded files
}
\keyword{config,}
\keyword{configuration,}
\keyword{filePath,}
\keyword{fileToParse}

