% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getURLcheckStatus}
\alias{deleteURLcheckStatus}
\alias{getURLcheckStatus}
\alias{postURLcheckStatus}
\alias{putURLcheckStatus}
\title{Get URL and check status}
\usage{
getURLcheckStatus(url, ..., requireJSON = FALSE)

postURLcheckStatus(url, postfields, ..., requireJSON = FALSE)

putURLcheckStatus(url, postfields, ..., requireJSON = FALSE)

deleteURLcheckStatus(url, ..., requireJSON = FALSE)
}
\arguments{
\item{url}{the url to get/post}

\item{...}{optional parameters passed to getURL}

\item{requireJSON}{boolean if errors should be thrown on JSON}

\item{postfields}{data sent to the server}
}
\description{
This is a wrapper for getURL that throws an error when the HTTP status is 400
or greater, or possibly when the response is html.
}
\details{
Checks the HTTP status and logs to racas.log as com.acas.sel if 400
 or greater. Setting requireJSON to \code{TRUE} will add a check for if the
 response is HTML (but not necessarily valid JSON). POST, PUT, and DELETE can
 be done with their own functions. In POST and PUT, the \code{postfields}
 will also be logged. Within \code{racas}, \code{postfields} is usually JSON.
}

