% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load_value_type_and_kinds}
\alias{load_value_type_and_kinds}
\title{Loads a set of default lsKinds per module}
\usage{
load_value_type_and_kinds(requiredModules = NA)
}
\arguments{
\item{requiredModules}{a character vector of modules to load see \link{modules}}
}
\value{
a list object of returned lsKinds
}
\description{
There are a set of required lsKinds that need to be loaded in modules. This function loads them all if given no input or a subset if provied
}

