% Generated by roxygen2 (4.0.1): do not edit by hand
\name{meltBatchCodes2}
\alias{meltBatchCodes2}
\title{Turns a batchCode column into rows in a long format}
\usage{
meltBatchCodes2(entityData)
}
\arguments{
\item{entityData}{a data frame with data}
}
\value{
A data frame with rows for all code values
}
\description{
Turns a batchCode column into rows in a long format
}
\details{
Does not work with data.table.
entityData must have columns "batchCode",
"tempStateId", "parentId", "tempId", "stateType", "stateKind"
If "batchCode" is missing, will return an empty data.frame.
publicData is always set to TRUE.
In longFormatSave.R.
For use with Tsv saves.
}

