% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/longFormatSave.R
\name{meltConcentrations}
\alias{meltConcentrations}
\title{Turns concentration columns into rows in a long format}
\usage{
meltConcentrations(entityData, entityKind = "treatmentGroup")
}
\arguments{
\item{entityData}{a data frame with data, must include rows "concentration" and "concentrationUnit"}

\item{entityKind}{the current acas entityKind (in racas::acasEntityHierarchyCamel)}
}
\description{
Turns concentration columns into rows in a long format
}

