% Generated by roxygen2 (4.0.0): do not edit by hand
\name{meltConcentrations}
\alias{meltConcentrations}
\title{Turns concentration columns into rows in a long format}
\usage{
meltConcentrations(entityData)
}
\arguments{
  \item{entityData}{a data frame with data, must include
  rows "concentration" and "concentrationUnit"}
}
\description{
Turns concentration columns into rows in a long format
}

