% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/longFormatSave.R
\name{meltTimes}
\alias{meltTimes}
\title{Turns time columns into rows in a long format}
\usage{
meltTimes(entityData, entityKind = "treatmentGroup")
}
\arguments{
\item{entityData}{a data frame with data, must include rows "time" and "timeUnit"}

\item{entityKind}{the current acas entityKind (in racas::acasEntityHierarchyCamel)}
}
\description{
Turns time columns into rows in a long format
}

