% Generated by roxygen2 (4.0.2): do not edit by hand
\name{postURLcheckStatus}
\alias{postURLcheckStatus}
\title{Post URL and check status}
\usage{
postURLcheckStatus(url, postfields, ...)
}
\arguments{
\item{url}{the url to get/post}

\item{postfields}{data sent to the server}

\item{...}{optional parameters passed to getURL}
}
\description{
This is similar to \link{getURLcheckStatus}, but does a POST, and the postfields are
logged in case of an error
}
\details{
checks the HTTP status and logs to racas.log as com.acas.sel if 400
 or greater
}

