% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/directDatabaseSave.R
\name{prepareTableForDD}
\alias{prepareTableForDD}
\title{Direct Database preparation}
\usage{
prepareTableForDD(entityData)
}
\arguments{
\item{entityData}{a data.frame or data.table, see source for columns.}
}
\value{
A data.table corrrectly formatted.
}
\description{
Prepares a data.frame or data.table for saving with
\code{\link{saveAgDataDD}} or similar. Coerces some columns to the correct
class and fills missing columns with NA.
}

