\name{readConfigFile}
\alias{readConfigFile}
\title{Read a configuration file to racas::applicationSettings}
\usage{
  readConfigFile(configLocation)
}
\arguments{
  \item{configLocation}{The location of the file to read}
}
\description{
  Read a configuration file to racas::applicationSettings
}
\keyword{applicationSettings,}
\keyword{config,}
\keyword{configuration,}
\keyword{configurationNode.js}

