% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readExcelOrCsv}
\alias{readExcelOrCsv}
\title{Reading of Excel or CSV}
\usage{
readExcelOrCsv(filePath, sheet = 1, header = FALSE)
}
\arguments{
\item{filePath}{The path the the file to read}

\item{sheet}{Sheet number of an Excel sheet}

\item{header}{Boolean to treat first row as column names}
}
\value{
a data.frame of the file
}
\description{
Read in xls, xlsx, csv, or txt (tab delimited) files and output a data frame of the information
}
\details{
Right now, this turns dates into "A_date_was_in_Excel_Date_format"
and does not treat headers as column names.
In fileRead.R
}

