% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/directDatabaseSave.R
\name{readTsvDataFileDD}
\alias{readTsvDataFileDD}
\title{Read tsv file for direct database load}
\usage{
readTsvDataFileDD(dataFilePath)
}
\arguments{
\item{dataFilePath}{path to tsv file}
}
\description{
Reads tsv data into a data.table for \code{\link{saveDataDirectDatabase}}
}

