% Generated by roxygen2 (4.0.2): do not edit by hand
\name{saveAcasEntity}
\alias{deleteAcasEntity}
\alias{getAcasEntities}
\alias{getAcasEntity}
\alias{saveAcasEntities}
\alias{saveAcasEntity}
\alias{updateAcasEntity}
\title{Save ACAS entities to the server}
\usage{
saveAcasEntity(entity, acasCategory,
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)

saveAcasEntities(entities, acasCategory,
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)

deleteAcasEntity(entity, acasCategory,
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)

updateAcasEntity(entity, acasCategory, acasLevel = "values",
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)

getAcasEntity()

getAcasEntities(acasCategory,
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{entity}{a single entity (a named list, becomes a JSON object)}

\item{acasCategory}{e.g. "experiments", "subjectlabels", etc.}

\item{lsServerURL}{url of ACAS server}

\item{entities}{a list of entities}
}
\value{
a list, sometimes empty
}
\description{
Save protocols, labels, experiments, etc.
}
\details{
\code{updateAcasEntities} replaces the entity that is at the URL with
  the one sent. Sub-entities (label, state, value) must have a parent object
}

