% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createHtmlSummary.R
\name{saveAnalysisResults}
\alias{saveAnalysisResults}
\title{Save the HTML summary into the experiment}
\usage{
saveAnalysisResults(experiment, hasError, htmlSummary, recordedBy = "none",
  dryRun = F, lsTransaction = NULL, testMode = FALSE)
}
\arguments{
\item{experiment}{An experiment list of lists}

\item{hasError}{A boolean in the analysis had an error}

\item{htmlSummary}{A string that is html}

\item{recordedBy}{current user (not used)}

\item{lsTransaction}{An integer that is the transaction id (not used)}

\item{testMode}{a boolean used for testing (not used)}
}
\value{
the htmlSummary
}
\description{
Saves a summary to the given experiment
}

