\name{saveAnalysisResults}
\alias{saveAnalysisResults}
\title{Save the HTML summary into the experiment}
\usage{
  saveAnalysisResults(experiment, hasError, htmlSummary,
    lsTransaction)
}
\arguments{
  \item{experiment}{An experiment list of lists}

  \item{hasError}{A boolean in the analysis had an error}

  \item{htmlSummary}{A string that is html}

  \item{lsTransaction}{An integer that is the transaction
  id}
}
\description{
  Saves a summary to the given experiment
}

