% Generated by roxygen2 (4.0.0): do not edit by hand
\name{saveAnalysisResults}
\alias{saveAnalysisResults}
\title{Save the HTML summary into the experiment}
\usage{
saveAnalysisResults(experiment, hasError, htmlSummary, lsTransaction = NULL,
  testMode = FALSE)
}
\arguments{
  \item{experiment}{An experiment list of lists}

  \item{hasError}{A boolean in the analysis had an error}

  \item{htmlSummary}{A string that is html}

  \item{lsTransaction}{An integer that is the transaction
  id}

  \item{testMode}{a boolean used for testing}
}
\description{
Saves a summary to the given experiment
}

