% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JSON_library.R
\name{saveContainer}
\alias{saveContainer}
\title{save container objects
Currently, this cannot accept labels and states}
\usage{
saveContainer(container,
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\description{
save container objects
Currently, this cannot accept labels and states
}

