% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getEntityIdsDD}
\alias{getEntityCodesBySqlDD}
\alias{getEntityCodesDD}
\alias{getEntityIdsDD}
\alias{getStateIdsDD}
\alias{getValueIdsDD}
\alias{saveAgDataDD}
\alias{saveEntitiesDD}
\alias{saveStatesDD}
\alias{saveSubjectDataDD}
\alias{saveTgDataDD}
\title{Helper functions for saving direct to database}
\usage{
getEntityIdsDD(conn, entityType, numberOfIds)

getStateIdsDD(conn, entityType, numberOfIds)

getValueIdsDD(conn, entityType, numberOfIds)

getEntityCodesBySqlDD(conn, entityType, numberOfCodes)

getEntityCodesDD(entityType, numberOfCodes)

saveAgDataDD(conn, inputDT, experimentId, lsTransactionId, recordedDate)

saveTgDataDD(conn, inputDT, ag_ids, lsTransactionId, recordedDate)

saveSubjectDataDD(conn, inputDT, tg_ids, lsTransactionId, recordedDate)

saveEntitiesDD(conn, entityType, inputDT)

saveStatesDD(conn, entityType, inputStatesDT)
}
\arguments{
\item{conn}{database connection object}

\item{entityType}{"ANALYSIS_GROUP", "TREATMENT_GROUP", or "SUBJECT"}

\item{numberOfIds}{number of ids to return}

\item{numberOfCodes}{number of codes to return}

\item{inputDT}{a data.table input. See source for required columns.}

\item{experimentId}{id of experiment}

\item{lsTransactionId}{integer of ls transaction id}

\item{recordedDate}{date of save, correctly string formatted for database}

\item{ag_ids}{data.table of analysis_group ids with columns tempParentId and id}

\item{tg_ids}{data.table of treatment_group ids with columns tempParentId and id}

\item{inputStatesDT}{a data.table input. See source for required columns.}
}
\description{
A set of functions used internally, expected to be used by \code{\link{saveDataDirectDatabase}}. Could be expanded later.
}

