\name{saveValueKinds}
\alias{saveValueKinds}
\title{Saves value kinds}
\usage{
  saveValueKinds(valueKinds, valueTypes, errorEnv = NULL)
}
\arguments{
  \item{valueKinds}{character vector of new valueKinds}

  \item{valueTypes}{character vector of valueTypes (e.g.
  c("stringValue", "numericValue"))}
}
\description{
  Saves value kinds with matching value types
}
\details{
  valueKinds must be new, and valueTypes must exist
}

