% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stopUser}
\alias{stopUser}
\title{Fatal error tracking}
\usage{
stopUser(message)
}
\arguments{
  \item{message}{The error message that the user should
  see}
}
\value{
Stops the function, and adds the class "userStop" to the
error object

All helpful errors should be thrown using \code{stopUser}.
Any error thrown using \code{stop} will be treated as an
internal error by the simple experiment loader.

When using this function, it is important to give it a
single string as an error message. This means using paste.
While \code{stop("text ", variable, " text")} is okay
syntax, you will get an error if you try
\code{stopUser("text ", variable, " text")}. Use paste0
instead: \code{stopUser(paste0("text ", variable, " text))}
(note that this syntax is also perfectly acceptable inside
\code{stop})
}
\description{
Declares an error to be of type "userStop", to distinguish
between errors we programmed into the system (e.g.
"Unrecognized scientist") and errors R gives (e.g. "object
not found")
}

