% Generated by roxygen2 (4.0.0): do not edit by hand
\name{tryCatch.W.E}
\alias{tryCatch.W.E}
\title{Error and warning trapping}
\usage{
tryCatch.W.E(expr)
}
\arguments{
\item{expr}{An expression which will have errors caught}
}
\value{
A list with value (either result or error) and warningList
}
\description{
Traps errors and warnings, creating lists of each
}

