% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tryCatchLog}
\alias{tryCatchLog}
\title{Error and warning trapping and logging}
\usage{
tryCatchLog(expr)
}
\arguments{
\item{expr}{An expression which will have errors caught}
}
\value{
A list with value, errorList, and warningList
}
\description{
Traps errors and warnings, logging them out
}
\details{
Don't nest this- it uses a global Messenger. \code{value} will be
\code{NULL} if there are any caught errors.
}

