% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JSON_library.R
\name{validateValueKindsFromDataFrame}
\alias{validateValueKindsFromDataFrame}
\title{validateValueKindsFromDataFrame}
\usage{
validateValueKindsFromDataFrame(typesAndKindsDataFrame,
  lsServerURL = racas::applicationSettings$client.service.persistence.fullpath)
}
\arguments{
\item{typesAndKindsDataFrame}{data.frame 2 columns lsType, lsKind}
}
\value{
a data table type names, kind names and full object kinds
}
\description{
Get a data table of type names and kind names plus the kind (full object) from a data frame of type names and kind names
}

